pyproject.toml
[tool.poetry]
name = "renewable energy"
version = "0.1.0"
description = ""
authors = ["Basma-G <127091566+Basma-Ghaffour@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^1.5.0"
owid = "^0.0.1"
pandas = "^2.2.2"
SQLAlchemy = "^2.0.32"
mysql-connector = "^2.2.9"
owid-catalog = "^0.3.11"
plotly-express = "^0.4.1"
plotly = "^5.23.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
mypy = "^1.11.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
